+++ Patches/ExtendedRadioPatches.cs
@@ -21,10 +21,22 @@ namespace ExtendedRadio.Patches
 
 		public static readonly string COUIBaseLocation = $"coui://{IconsResourceKey}";
 
+		static readonly string CustomRadioPath = Path.Combine(Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location), "CustomRadio");
+		public static readonly string CustomRadiosPath = Path.Combine(Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location), "CustomRadios");
+		static readonly string PathToParent = Directory.GetParent(Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location)).FullName;
+		public static readonly string PathToMods = Path.Combine(PathToParent,"ExtendedRadio_mods");
+		public static readonly string CustomRadioFolderPlugins = Path.Combine(PathToMods,"CustomRadios");
+
 		static void Prefix(GameManager __instance)
 		{		
 
-			Directory.CreateDirectory(Path.Combine(Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location), "CustomRadio"));
+			List<string> pathToIconToLoad = [Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location)];
+
+			if(Directory.Exists(CustomRadioPath)) {
+				Directory.Move(CustomRadioPath, CustomRadiosPath);
+			}
+
+			Directory.CreateDirectory(CustomRadiosPath);
 
 			string resources = Path.Combine(Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location), "resources");
 
@@ -33,10 +45,10 @@ namespace ExtendedRadio.Patches
 				File.Move(Path.Combine(Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location), "DefaultIcon.svg"), Path.Combine(resources , "DefaultIcon.svg"));
 			}
 
-			string CustomRadioFolderPlugins = Path.Combine(Directory.GetParent(Assembly.GetExecutingAssembly().Location).Name,"CustomRadio");
-
+			Debug.Log(CustomRadioFolderPlugins);
 			if(Directory.Exists(CustomRadioFolderPlugins)) {
 				ExtendedRadio.RegisterCustomRadioDirectory(CustomRadioFolderPlugins);
+				pathToIconToLoad.Add(PathToMods);
 			}
 
 			var gameUIResourceHandler = (GameUIResourceHandler)GameManager.instance.userInterface.view.uiSystem.resourceHandler;
@@ -48,13 +60,17 @@ namespace ExtendedRadio.Patches
 			}
 			
 			gameUIResourceHandler.HostLocationsMap.Add(
-				IconsResourceKey,
-				[
-					Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location)
-				]
+				IconsResourceKey, pathToIconToLoad
+
 			);
 		}
 	}
+
+				// 	[
+				// 	Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location),
+
+
+				// ]
 
 	[HarmonyPatch(typeof( Radio ), "LoadRadio")]
 	class Radio_LoadRadio {