+++ CHANGELOG.md
@@ -1,5 +1,9 @@
-# DEV
-- Add Two new radio mod format that support all the possible customisation
+# 0.3.0-pre-9
+- Add: A new folder can be used for Custom Radio outside the mod folder to avoid is deletion when the mod update on Thunderstore Mod Manager. (`BepInEx/plugins/ExtendedRadio_mods/CustomRadios`)
+- Add: A complete new way to create radio, that support a lot of customization. (The support for the 0.2.x way, is always in the mod)
+- Add: The API is finally there, you can now create radio using mod.
+- Update: The wiki is update for the 0.3.0pre-5 version of the mod.
+- Fixed: Bug.
 
 # 0.2.1
 - Fixed: Changelog